feat: add serde serialization for validator_pubkey in User struct (#1496)
This pull request makes a small change to the `User` struct in
`smartcontract/programs/doublezero-serviceability/src/state/user.rs` to
improve serialization and deserialization of the `validator_pubkey`
field when the `serde` feature is enabled.

* Serialization improvements:
* Added custom `serde` attributes to the `validator_pubkey` field to
ensure it is serialized as a string and deserialized correctly using
helper functions from `doublezero_program_common::serializer`.

## Testing Verification
* Show evidence of testing the change

Co-authored-by: Juan Olveira <jolveira@gmail.com>