Add mock swap program (malbeclabs/doublezero-solana#51)
Add a mock swap program to prove out withdraw SOL integration and sweep
distribution tokens integration with an example swap program.

This PR fixes the sweep-distribution-tokens and distribute-rewards tests
by adding mock-buy-SOL calls, which are used to adjust the Journal
balances when SOL is converted to 2Z tokens. And now both feature sets
("development" and not "development") can be tested.

This PR also removes the 2z-cli crate since it now lives in the
[here](https://github.com/doublezerofoundation/doublezero-offchain/tree/main/crates/2z-cli).

Closes https://github.com/malbeclabs/doublezero/issues/777.
Closes https://github.com/malbeclabs/doublezero/issues/779.
Closes https://github.com/malbeclabs/doublezero/issues/1038.