validator-debt: add checks after writing to ledger (malbeclabs/doublezero-offchain#95)
This PR adds a couple of checks to ledger operations. First, it checks
to see if the blockhash for the newly computed entry to ensure that it's
newer than the existing blockhash (if it exists). If the record doesn't
exist it creates the record and then exits; the next run since the
record exists it will ensure that the record details match and then will
write the transaction on solana.

Closes https://github.com/malbeclabs/doublezero/issues/1648

## Testing Evidence 

```
➜  validator-debt git:(add_ledger_checks) ✗ ../../target/release/doublezero-solana-validator-debt calculate-validator-debt --dz-ledger-url $LEDGER_RPC -u $SOLANA_RPC --epoch 91 --dry-run

~snip~

wrote 84 bytes for blockhash 87ST7sby6USb6EsC3mn8YUZuHY6xgDZs2tFof9uE4UqN
Writing total debt to solana
Simulated program logs:
  Program dzrevZC94tBLwuHw1dyynZxaXTWyp7yocsinyEVPtt4 invoke [1]
  Program log: Configure distribution debt
  Program log: DZ epoch: 91
  Program log: Set total_solana_validators: 1
  Program log: Set total_solana_validator_debt: 495
  Program log: Set solana_validator_debt_merkle_root: 2Ea4FHVHn8mFsEaA6SMTZ5YUb7yvc7GAYqfCVK6tgzsy
  Program dzrevZC94tBLwuHw1dyynZxaXTWyp7yocsinyEVPtt4 consumed 4405 of 200000 compute units
  Program dzrevZC94tBLwuHw1dyynZxaXTWyp7yocsinyEVPtt4 success
Validator rewards for solana epoch 838 and validator debt for DoubleZero epoch 91:
 validator_pubkey                            | total_debt | total_rewards | block_base_rewards | block_priority_rewards | inflation_rewards | jito_rewards
 va1i6T6vTcijrCz6G8r89H6igKjwkLfF6g5fnpvZu1b | 495        | 122806394101  | 4957500            | 5000                   | 122801431601      | 0
```