crates: add client-tools and admin CLIs (malbeclabs/doublezero-solana#65)
doublezero-solana-client-tools can be helpful for various CLIs (namely
the doublezero-solana CLI in the offchain repo).

The admin CLIs are meant to replace the admin subcommands in the
doublezero-solana CLI so we do not ship the admin commands with a
user-facing interface.

Also uptick Solana dependencies to 2.3.9 and clean up unused
dependencies.