api: add web service for token supply stats (#1687)
## Summary of Changes
We need to provide folks an http endpoint to get current total tokens in
circulation. This is a small webservice that does the needful. It pulls
the current supply from solana mainnet via the getSupply RPC to factor
in burn/inflation changes and daily lockup information from an s3 bucket
then does math.

## Testing Verification
```
➜ ~ S3_ENDPOINT=<redacted> S3_ACCESS_KEY=<redacted> S3_ACCESS_SECRET=<redacted> ./api_server -use-s3
{"time":"2025-09-19T16:59:39.143885-04:00","level":"INFO","msg":"API server starting","address":":8080"}

➜  ~ curl localhost:8080/supply
3471417500.0%
```