revenue-distribution: add initialize distribution grace period (malbeclabs/doublezero-solana#78)
This change adds an extra layer of protection against the debt
calculator from initializing distributions too quickly. The system
trusts that the debt accountant is staying in sync with the DoubleZero
Ledger network. But the new initialization grace period safeguard will
allow ample time to address a misbehaving debt accountant.

We expect to configure this new parameter to 24 or 36 hours.

This PR also changes a couple of configuration types:
- calculation grace period seconds (u32) to MINUTES (u16)
- minimum epochs to finalize rewards u16 to u8 because anything >255
feels excessive

Closes https://github.com/malbeclabs/doublezero/issues/1785.