validator-debt: integrate slack notifications (malbeclabs/doublezero-offchain#161)
## Summary of Changes

This PR adds a `slack-notifier` crate which allows the various CLIs to
post output to slack. Currently it' s only for the validator-debt crate
but can be used by any of the others. After a `calculate-debt`,
`finalize-distribution`, or `pay-solana-validator-debt` command is
issued, a slack message is generated.

## Testing Verification 

One can see the testing output in the [tmp
channel](https://malbeclabs.slack.com/archives/C09LES1Q127)