validator-debt: parse program logs, attach exported csv to slack msg (malbeclabs/doublezero-offchain#163)
## Summary of Changes

This PR adds CSVs to slack messages and parses the program logs returned
from solana to provide more detail about the success or failure of a
transaction. It reorganizes some of the common slack functions into
`slack.rs` so that they might be more easily reused.

Closes https://github.com/malbeclabs/doublezero/issues/1972
Closes https://github.com/malbeclabs/doublezero/issues/1974

## Testing Verification 
Testing evidence can be found in
[#tmp-validator-debt](https://malbeclabs.slack.com/archives/C09LES1Q127)
showing both the updated message format and the CSV file URL.