feat(scheduler): Automatically finalize distribution (malbeclabs/doublezero-offchain#200)
## Summary of Changes

This PR updates the calculate_distribution worker to add in a counter
that increments each time the calculate_distribution function
successfully completes. When the counter reaches three, it moves to the
finalize_distribution callback where the finalize_distribution NIF is
called. If any of the calls fail or the computed_debt differs from the
ledger debt record, the process is stopped and the counter is reset to
zero.

Note that only the third successful run is posted to slack.

Closes https://github.com/malbeclabs/doublezero/issues/2028

## Testing Verification

* This was run and posted to slack as evidenced
[here](https://malbeclabs.slack.com/archives/C09LES1Q127/p1763585298695259)