fix(sentinel): improve previous epoch slot evaluation (malbeclabs/doublezero-offchain#213)
## Summary of Changes
When using a local validator, the leader schedule check will always fail
because the previous epoch slot iterator will always return `None` since
the current slot will always be less than 432,000 slots (the number of
slots in an epoch). So if we were to test the Sentinel in a local
environment, it never has a chance of granting access. This PR fixes
this iterator evaluation.

This change also makes `PreviousEpochSlots` private since nothing else
would use it besides this submodule.

Closes https://github.com/malbeclabs/doublezero/issues/2306.

## Testing Verification
Unit tests still pass.