feat(validator-debt): better summarized debt  (malbeclabs/doublezero-offchain#218)
## Summary of Changes
This PR adds in more details for when debt collection is posted to slack
and also summarizes debt collection for all epochs. It moves some of the
debt-related slack composition to the validator-debt crate so we can
reuse the existing structs instead of a many-arity function. It also
adds total debt, total paid, and insufficient payment counts to the pay
debt GenServer which are then posted to slack at the end of a debt
collection run.

Closes https://github.com/malbeclabs/doublezero/issues/2256.

## Testing Verification
* This was tested locally in a private slack channel:

<img width="1269" height="702" alt="Screenshot 2025-12-02 at 17 19 46"
src="https://github.com/user-attachments/assets/a8f75951-01e0-4ba4-a4f6-9d8cf2657dab"
/>