telemetry: add serviceability data to flow-enricher (#2448)
## Summary of Changes
This adds the periodic fetching and caching of serviceability data to
the flow enricher which will allow us to add metadata to incoming flow
data. Flow annotation will be done in an upcoming PR. Without this
change, flow data is not as useful as it could be since we wouldn't know
user information, exchange, region, exporting device for the flow.

## Testing Verification
There are unit tests attached to demonstrate initial/periodic fetching
of serviceability data using an interface stub.

closes #2447