feat(solana-cli): add `--fund-outstanding-debt` to `revenue-distribution validator-deposit` (malbeclabs/doublezero-offchain#231)
## Summary of Changes
This change adds an argument to the `revenue-distribution
validator-deposit` command called `--fund-outstanding-debt`, which
fetches all debt for a given Solana validator node ID and transfers SOL
based on the outstanding amount, which is the total debt that has been
unpaid minus the deposit account balance).

Also added a helper method in the `doublezero-solana-validator-debts`
crate that both this command and the `revenue-distribution fetch
validator-debts` command leverage.

Closes https://github.com/malbeclabs/doublezero/issues/2247.

## Testing Verification
CLI calls in local validator shell script test work.