telemetry: add serviceability annotator to flow-enricher (#2464)
## Summary of Changes
This adds a flow annotator to the enricher that allows for flows to
contain serviceability metadata, namely:
- source device
- destination device
- source location
- destination location
- source exchange
- destination exchange

This is done by looking at the source/destination IP in the flow record
and attempting to find a match against user DZ IPs onchain, which then
allow for device/location/exchange lookups by pubkey.

## Testing Verification
There are unit tests for the annotator as well as the integration test
has been updated.