actions: cancel existing e2e run on new trigger (#2467)
## Summary of Changes
This causes e2e workflows to be cancelled if an existing PR or branch
triggers a second run of the workflow. If you push new commits to an
open PR, it can create a backlog of e2e tests to run on old commits,
which is useless.

## Testing Verification
This change was pushed without updating the changelog but still kicked
off an e2e run. When pushing to this branch with the updated changelog,
the old e2e run was immediately cancelled:
https://github.com/malbeclabs/doublezero/actions/runs/20343856767