telemetry: add expanded flow sample parsing (#2501)
## Summary of Changes
This adds parsing of expanded format flow samples to the flow enricher.
We were previously only parsing standard format samples but testnet
devices use subinterfaces which forces the expanded format to be used on
some nodes.

## Testing Verification
There is a test attached that reads in a pcap of the expanded format
payload and verifies the payload is decoded as expected.