telemetry: make flow-enricher table configurable (#2515)
## Summary of Changes
This adds support for setting the clickhouse table name in the
flow-enricher. The table name was hardcoded as "flows" which doesn't
allow us to set a different table per environment which prevents us from
testing schema changes so we don't impact mainnet.

## Testing Verification
The flow-enricher integration test was configured to use the new
`WithClickhouseTable` option with a new table name. If the test passes,
flows were inserted into the correct table and read back out correctly.